refactor to raylib

+++ lib/rngs/b_rand.c
@@ -13,7 +13,7 @@ uint64_t splitmix64(uint64_t seed) {
   return seed;
 }
 //99% sure this is compiler optimized anyways
-float normalize64(uint64_t seed) { return (float)seed / (float)UINT64_MAX; }
+double normalize64(uint64_t seed) { return (float)seed / (float)UINT64_MAX; }
 
 uint64_t xorshift64(uint64_t seed) {
   uint64_t state = seed;
